updating with 18.2.10 assets (#1242)
* updating  for api_guide.tgz release 18.2.10

* updating  for swagger.tgz release 18.2.10

* updating go/clients for avi_go_sdk.tgz release 18.2.10

* updating go/models for avi_go_sdk.tgz release 18.2.10

* updated version

+++ go/models/controller_license.go
@@ -18,7 +18,7 @@ type ControllerLicense struct {
 	BurstCores *int32 `json:"burst_cores,omitempty"`
 
 	// Number of Service Engine cores in non-container clouds.
-	Cores *int32 `json:"cores,omitempty"`
+	Cores *float64 `json:"cores,omitempty"`
 
 	// customer_name of ControllerLicense.
 	// Required: true
